+++ src/components/profile/profile-list.tsx
@@ -1,0 +1,36 @@
+import { ScrollableContent } from '@/components/common/scrollable-content'
+import { useTranslations } from 'next-intl'
+import { ProfileCard, ProfileWithStats } from './profile-card'
+
+interface ProfileListProps {
+  profiles: ProfileWithStats[]
+  isLoading: boolean
+  router: any
+}
+
+export const ProfileList = ({
+  profiles,
+  isLoading,
+  router,
+}: ProfileListProps) => {
+  const t = useTranslations()
+
+  return (
+    <ScrollableContent
+      isLoading={isLoading}
+      isEmpty={profiles.length === 0}
+      loadingText={t('profile_info.fetching_profiles')}
+      emptyText={t('profile_info.no_profile_found')}
+    >
+      <div className="divide-y divide-green-800/30">
+        {profiles.map((profile) => (
+          <ProfileCard
+            key={`${profile.profile.username}-${profile.namespace?.name}`}
+            profile={profile}
+            router={router}
+          />
+        ))}
+      </div>
+    </ScrollableContent>
+  )
+}
